refactor: teams endpoints
