+++ src/main/java/ecostruxure/rate/calculator/bll/service/CurrencyService.java
@@ -1,16 +1,10 @@
 package ecostruxure.rate.calculator.bll.service;
 
-import com.opencsv.CSVReader;
-import com.opencsv.exceptions.CsvException;
 import ecostruxure.rate.calculator.be.Currency;
-import ecostruxure.rate.calculator.bll.MissingCurrenciesException;
-import ecostruxure.rate.calculator.dal.interfaces.ICurrencyRepository;
+import ecostruxure.rate.calculator.dal.ICurrencyRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.io.FileReader;
-import java.io.IOException;
-import java.math.BigDecimal;
 import java.util.*;
 
 @Service
