+++ src/main/java/ecostruxure/rate/calculator/dal/IGeographyRepository.java
@@ -1,4 +1,4 @@
-package ecostruxure.rate.calculator.dal.interfaces;
+package ecostruxure.rate.calculator.dal;
 
 import ecostruxure.rate.calculator.be.Geography;
 import org.springframework.data.repository.CrudRepository;
