+++ src/main/java/ecostruxure/rate/calculator/dal/IProfileRepository.java
@@ -1,4 +1,4 @@
-package ecostruxure.rate.calculator.dal.interfaces;
+package ecostruxure.rate.calculator.dal;
 
 import ecostruxure.rate.calculator.be.Profile;
 import org.springframework.data.repository.CrudRepository;
