+++ src/main/java/ecostruxure/rate/calculator/dal/IProjectTeamRepository.java
@@ -1,4 +1,4 @@
-package ecostruxure.rate.calculator.dal.interfaces;
+package ecostruxure.rate.calculator.dal;
 
 import ecostruxure.rate.calculator.be.ProjectTeam;
 import org.springframework.data.jpa.repository.support.SimpleJpaRepository;
